Remove the old parser
