+++ crates/cel-fork/cel/Cargo.toml
@@ -10,8 +10,6 @@ license = "MIT"
 categories = ["compilers"]
 
 [dependencies]
-antlr4rust = "0.5.2"
-lazy_static = "1.5.0"
 nom = "7.1.3"
 chrono = { version = "0.4", default-features = false, features = ["alloc", "serde", "std"] }
 regex = { version = "1.10.5"}
