--- crates/cel-fork/cel/src/parser/gen/CEL.tokens
@@ -1,65 +1,0 @@
-EQUALS=1
-NOT_EQUALS=2
-IN=3
-LESS=4
-LESS_EQUALS=5
-GREATER_EQUALS=6
-GREATER=7
-LOGICAL_AND=8
-LOGICAL_OR=9
-LBRACKET=10
-RPRACKET=11
-LBRACE=12
-RBRACE=13
-LPAREN=14
-RPAREN=15
-DOT=16
-COMMA=17
-MINUS=18
-EXCLAM=19
-QUESTIONMARK=20
-COLON=21
-PLUS=22
-STAR=23
-SLASH=24
-PERCENT=25
-CEL_TRUE=26
-CEL_FALSE=27
-NUL=28
-WHITESPACE=29
-COMMENT=30
-NUM_FLOAT=31
-NUM_INT=32
-NUM_UINT=33
-STRING=34
-BYTES=35
-IDENTIFIER=36
-ESC_IDENTIFIER=37
-'=='=1
-'!='=2
-'in'=3
-'<'=4
-'<='=5
-'>='=6
-'>'=7
-'&&'=8
-'||'=9
-'['=10
-']'=11
-'{'=12
-'}'=13
-'('=14
-')'=15
-'.'=16
-','=17
-'-'=18
-'!'=19
-'?'=20
-':'=21
-'+'=22
-'*'=23
-'/'=24
-'%'=25
-'true'=26
-'false'=27
-'null'=28
